+++ frontend/src/pages/Campgrounds/subscreen/components/ReviewModal.jsx
@@ -1,67 +1,21 @@
-import { Alert, Button, Group, Input, LoadingOverlay, Modal, Paper, Rating, Textarea, Title } from "@mantine/core";
+import { Alert, Button, Group, Modal, Paper, Title } from "@mantine/core";
 import { useDisclosure } from "@mantine/hooks";
-import { useForm } from '@mantine/form';
 import { IconAlertCircle } from "@tabler/icons-react";
-import { useNavigate } from "react-router-dom";
-import axios from "axios";
 import { useContext } from "react";
 import { AuthContext } from "../../../../auth/AuthContext";
 import LoginForm from "../../../Login/components/LoginForm";
+import ReviewForm from "./ReviewForm";
 
-export default function ({ campgroundId }) {
+export default function () {
     const [opened, { open, close }] = useDisclosure(false);
-    const [visible, { toggle }] = useDisclosure(false);
-    const navigate = useNavigate();
     const { user } = useContext(AuthContext);
-
-    const form = useForm({
-        initialValues: {
-            rating: 0,
-            body: '',
-        },
-
-        validate: {
-            rating: (value) => (value === 0 ? 'Rating is required' : null),
-            body: (value) => (value.length < 10 ? 'Review body must be at least 10 characters long' : null),
-        },
-    });
-
-    const postReview = async (review) => {
-        await axios.post(`http://localhost:4000/campgrounds/${campgroundId}/reviews`, { review })
-            .then(() => {
-                navigate(0);
-            })
-            .catch(error => {
-                console.error(error);
-            })
-    }
 
     return (
         <>
             <Modal opened={opened} onClose={close} title={<Title order={5}>Submit a review</Title>} radius="md" centered>
                 <Paper p={14} radius="md" miw={400}>
                     {user ? (
-                        <form onSubmit={form.onSubmit(postReview)}>
-                            <Group position="center">
-                                <Input.Wrapper size="md" {...form.getInputProps('rating')}>
-                                    <Rating
-                                        m={5}
-                                        size="xl"
-                                        {...form.getInputProps('rating')}
-                                    />
-                                </Input.Wrapper>
-                            </Group>
-                            <Textarea
-                                my={20}
-                                size="md"
-                                placeholder="Share your experience in this campground"
-                                {...form.getInputProps('body')}
-                            />
-                            <Group position="right" mt="md">
-                                <Button type="button" onClick={close} variant="outline">Cancel</Button>
-                                <Button type="submit" onClick={toggle} disabled={!form.isDirty('rating') || !form.isDirty('body')}>Submit review</Button>
-                            </Group>
-                        </form>
+                        <ReviewForm />
                     ) : (
                         <>
                             <Alert icon={<IconAlertCircle size="1rem" />} title="">
@@ -70,7 +24,6 @@ export default function ({ campgroundId }) {
                             <LoginForm />
                         </>
                     )}
-                    <LoadingOverlay visible={visible} overlayBlur={2} />
                 </Paper>
             </Modal>
             <Group my={25}>
